pyproject.toml
[project]
name='EvoRube'
version='0.0.1'
requires-python='>=3.10'
dependencies=[
  'numpy>=1.26',
  'pyarrow>=15',
  'duckdb>=0.10',
  'PyYAML>=6']

[build-system]
requires=['setuptools','wheel']

[tool.setuptools]
packages = ["rubiks"]